Move lookup up to LF::prepare, avoid some MT creation, minor efficiency tweaks
